Finished Fruit Microservice

+++ fruit-microservice/fruit/src/main/java/com/bites/fruit/repository/IFruitRepository.java
@@ -4,6 +4,6 @@ import com.bites.fruit.model.Fruit;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IFruitRepository extends JpaRepository<Fruit, Long> {
+public interface IFruitRepository extends JpaRepository<Fruit, Integer> {
     Fruit findByName(String name);
 }
